Do not interpret Enum members called 'description' as description properties
This is a workaround for `TypeError`s being raised when initialising schemas with
Enum members named `description` or `deprecation_reason`.

It is still not ideal as you can't have an enum with both a "description" member
and a description property. Fixing this would require some more thought though.
Potentially we could allow something like `__description__` as an alternative way of
naming the description property in cases where it conflicts with one of the members.

Fixes #1321